@@ -16,47 +16,44 @@
 
 /// <reference path="../../typings/tsd.d.ts" />
 
-import CaveatPacket = require('./CaveatPacket');
-import CaveatPacketType = require('./CaveatPacketType');
-import Macaroon = require('./Macaroon');
-import MacaroonsConstants = require('./MacaroonsConstants');
-import BufferTools = require('./BufferTools');
-import MacaroonsDeSerializer = require('./MacaroonsDeSerializer');
-import CryptoTools = require('./CryptoTools');
-
-export = MacaroonsVerifier;
+import BufferTools from './BufferTools';
+import CaveatPacket from './CaveatPacket';
+import {CaveatPacketType} from './CaveatPacketType';
+import CryptoTools from './CryptoTools';
+import Macaroon from './Macaroon';
+import MacaroonsConstants from './MacaroonsConstants';
 
 /**
  * Used to verify Macaroons
  */
-class MacaroonsVerifier {
-
-  "use strict";
+export default class MacaroonsVerifier {
 
   private predicates:string[] = [];
   private boundMacaroons:Macaroon[] = [];
-  private generalCaveatVerifiers:GeneralCaveatVerifier[] = [];
+  private generalCaveatVerifiers:IGeneralCaveatVerifier[] = [];
   private macaroon:Macaroon;
 
   constructor(macaroon:Macaroon) {
     this.macaroon = macaroon;
   }
 
-  /**
-   * @param secret string this secret will be enhanced, in case it's shorter than {@link MacaroonsConstants.MACAROON_SUGGESTED_SECRET_LENGTH}
-   * @throws Error if macaroon isn't valid
-   */
-  public assertIsValid(secret:string):void;
+  private static containsElement(elements:string[], anElement:string):boolean {
+    if(elements !== null && elements.filter(e => e === anElement).length > 0) {
+      return true
+    }
+
+    return false;
+ }
+
   /**
    * @param secret a Buffer, that will be used, a minimum length of {@link MacaroonsConstants.MACAROON_SUGGESTED_SECRET_LENGTH} is highly recommended
    * @throws Error if macaroon isn't valid
    */
-  public assertIsValid(secret:Buffer):void;
-  public assertIsValid(secret:any):void {
-    var secretBuffer = (secret instanceof Buffer) ? secret : CryptoTools.generate_derived_key(secret);
-    var result = this.isValid_verify_raw(this.macaroon, secretBuffer);
+  public assertIsValid(secret:string|Buffer):void {
+    const secretBuffer = (secret instanceof Buffer) ? secret : CryptoTools.generate_derived_key(secret);
+    const result = this.isValid_verify_raw(this.macaroon, secretBuffer);
     if (result.fail) {
-      var msg = result.failMessage != null ? result.failMessage : "This macaroon isn't valid.";
+      const msg = result.failMessage !== undefined ? result.failMessage : "This macaroon isn't valid.";
       throw new Error(msg);
     }
   }
@@ -65,14 +62,9 @@ class MacaroonsVerifier {
    * @param secret string this secret will be enhanced, in case it's shorter than {@link MacaroonsConstants.MACAROON_SUGGESTED_SECRET_LENGTH}
    * @return true/false if the macaroon is valid
    */
-  public isValid(secret:string):boolean;
-  /**
-   * @param secret a Buffer, that will be used, a minimum length of {@link MacaroonsConstants.MACAROON_SUGGESTED_SECRET_LENGTH} is highly recommended
-   * @return true/false if the macaroon is valid
-   */
-  public isValid(secret:Buffer):boolean;
-  public isValid(secret:any):boolean {
-    var secretBuffer = (secret instanceof Buffer) ? secret : CryptoTools.generate_derived_key(secret);
+  public isValid(secret:string|Buffer):boolean {
+    const secretBuffer =  CryptoTools.generate_derived_key(secret.toString());
+    
     return !this.isValid_verify_raw(this.macaroon, secretBuffer).fail;
   }
 
@@ -91,6 +83,7 @@ class MacaroonsVerifier {
     if (caveat) {
       this.predicates.push(caveat);
     }
+
     return this;
   }
 
@@ -104,6 +97,7 @@ class MacaroonsVerifier {
     if (preparedMacaroon) {
       this.boundMacaroons.push(preparedMacaroon);
     }
+
     return this;
   }
 
@@ -124,132 +118,129 @@ class MacaroonsVerifier {
     if (generalVerifier) {
       this.generalCaveatVerifiers.push(generalVerifier);
     }
+
     return this;
   }
 
   private isValid_verify_raw(M:Macaroon, secret:Buffer):VerificationResult {
-    var vresult = this.macaroon_verify_inner(M, secret);
+    let vresult = this.macaroon_verify_inner(M, secret);
     if (!vresult.fail) {
       vresult.fail = !BufferTools.equals(vresult.csig, this.macaroon.signatureBuffer);
       if (vresult.fail) {
         vresult = new VerificationResult("Verification failed. Signature doesn't match. Maybe the key was wrong OR some caveats aren't satisfied.");
       }
     }
+    
     return vresult;
   }
 
   private macaroon_verify_inner(M:Macaroon, key:Buffer):VerificationResult {
-    var csig:Buffer = CryptoTools.macaroon_hmac(key, M.identifier);
-    if (M.caveatPackets != null) {
-      var caveatPackets = M.caveatPackets;
-      for (var i = 0; i < caveatPackets.length; i++) {
-        var caveat = caveatPackets[i];
-        if (caveat == null) continue;
-        if (caveat.type == CaveatPacketType.cl) continue;
-        if (!(caveat.type == CaveatPacketType.cid && caveatPackets[Math.min(i + 1, caveatPackets.length - 1)].type == CaveatPacketType.vid)) {
+    let csig:Buffer = CryptoTools.macaroon_hmac(key, M.identifier);
+
+    if (M.caveatPackets !== undefined) {
+      const caveatPackets:CaveatPacket[] = M.caveatPackets;
+      for (let i = 0; i < caveatPackets.length; i++) {
+        const caveat = caveatPackets[i];
+        if (caveat === undefined) { continue; }
+        if (caveat.type === CaveatPacketType.cl) { continue; }
+        if (!(caveat.type === CaveatPacketType.cid && caveatPackets[Math.min(i + 1, caveatPackets.length - 1)].type === CaveatPacketType.vid)) {
           if (MacaroonsVerifier.containsElement(this.predicates, caveat.getValueAsText()) || this.verifiesGeneral(caveat.getValueAsText())) {
             csig = CryptoTools.macaroon_hmac(csig, caveat.rawValue);
           }
         } else {
           i++;
-          var caveat_vid = caveatPackets[i];
-          var boundMacaroon = this.findBoundMacaroon(caveat.getValueAsText());
-          if (boundMacaroon == null) {
-            var msg = "Couldn't verify 3rd party macaroon, because no discharged macaroon was provided to the verifier.";
+          const caveatVid = caveatPackets[i];
+          const boundMacaroon = this.findBoundMacaroon(caveat.getValueAsText());
+          
+          if (boundMacaroon === undefined) {
+            const msg = "Couldn't verify 3rd party macaroon, because no discharged macaroon was provided to the verifier.";
+
             return new VerificationResult(msg);
           }
-          if (!this.macaroon_verify_inner_3rd(boundMacaroon, caveat_vid, csig)) {
-            var msg = "Couldn't verify 3rd party macaroon, identifier= " + boundMacaroon.identifier;
+          if (!this.macaroon_verify_inner_3rd(boundMacaroon, caveatVid, csig)) {
+            const msg = `Couldn't verify 3rd party macaroon, identifier= ${boundMacaroon.identifier}`;
+
             return new VerificationResult(msg);
           }
-          var data = caveat.rawValue;
-          var vdata = caveat_vid.rawValue;
+          const data = caveat.rawValue;
+          const vdata = caveatVid.rawValue;
           csig = CryptoTools.macaroon_hash2(csig, vdata, data);
         }
       }
     }
+
     return new VerificationResult(csig);
   }
 
   private macaroon_verify_inner_3rd(M:Macaroon, C:CaveatPacket, sig:Buffer):boolean {
-    if (!M) return false;
-    var enc_plaintext = new Buffer(MacaroonsConstants.MACAROON_SECRET_TEXT_ZERO_BYTES + MacaroonsConstants.MACAROON_HASH_BYTES);
-    var enc_ciphertext = new Buffer(MacaroonsConstants.MACAROON_HASH_BYTES + MacaroonsConstants.SECRET_BOX_OVERHEAD);
-    enc_plaintext.fill(0);
-    enc_ciphertext.fill(0);
+    if (!M) { return false; }
+    let encPlaintext = new Buffer(MacaroonsConstants.MACAROON_SECRET_TEXT_ZERO_BYTES + MacaroonsConstants.MACAROON_HASH_BYTES);
+    const encCipherText = new Buffer(MacaroonsConstants.MACAROON_HASH_BYTES + MacaroonsConstants.SECRET_BOX_OVERHEAD);
+    encPlaintext.fill(0);
+    encCipherText.fill(0);
 
-    var vid_data = C.rawValue;
-    //assert vid_data.length == VID_NONCE_KEY_SZ;
+    const vidData = C.rawValue;
+    //assert vidData.length == VID_NONCE_KEY_SZ;
     /**
      * the nonce is in the first MACAROON_SECRET_NONCE_BYTES
      * of the vid; the ciphertext is in the rest of it.
      */
-    var enc_nonce = new Buffer(MacaroonsConstants.MACAROON_SECRET_NONCE_BYTES);
-    vid_data.copy(enc_nonce, 0, 0, enc_nonce.length);
+    const encNonce = new Buffer(MacaroonsConstants.MACAROON_SECRET_NONCE_BYTES);
+    vidData.copy(encNonce, 0, 0, encNonce.length);
 
-    /* fill in the ciphertext */
-    vid_data.copy(enc_ciphertext, 0, MacaroonsConstants.MACAROON_SECRET_NONCE_BYTES, MacaroonsConstants.MACAROON_SECRET_NONCE_BYTES + vid_data.length - MacaroonsConstants.MACAROON_SECRET_NONCE_BYTES);
+    // fill in the ciphertext 
+    vidData.copy(encCipherText, 0, MacaroonsConstants.MACAROON_SECRET_NONCE_BYTES, MacaroonsConstants.MACAROON_SECRET_NONCE_BYTES + vidData.length - MacaroonsConstants.MACAROON_SECRET_NONCE_BYTES);
     try {
-      var enc_plaintext = CryptoTools.macaroon_secretbox_open(sig, enc_nonce, enc_ciphertext);
+      encPlaintext = CryptoTools.macaroon_secretbox_open(sig, encNonce, encCipherText);
     }
     catch (error) {
       if (/Cipher bytes fail verification/.test(error.message)) {
         return false;
       } else {
-        throw new Error("Error while deciphering 3rd party caveat, msg=" + error);
+        throw new Error(`Error while deciphering 3rd party caveat, msg=${error}`);
       }
     }
-    var key = new Buffer(MacaroonsConstants.MACAROON_HASH_BYTES);
+    const key = new Buffer(MacaroonsConstants.MACAROON_HASH_BYTES);
     key.fill(0);
-    enc_plaintext.copy(key, 0, 0, MacaroonsConstants.MACAROON_HASH_BYTES);
-    var vresult = this.macaroon_verify_inner(M, key);
+    encPlaintext.copy(key, 0, 0, MacaroonsConstants.MACAROON_HASH_BYTES);
+    const vresult = this.macaroon_verify_inner(M, key);
 
-    var data = this.macaroon.signatureBuffer;
-    var csig = CryptoTools.macaroon_bind(data, vresult.csig);
+    const data = this.macaroon.signatureBuffer;
+    const csig = CryptoTools.macaroon_bind(data, vresult.csig);
 
     return BufferTools.equals(csig, M.signatureBuffer);
   }
 
   private findBoundMacaroon(identifier:string):Macaroon {
-    for (var i = 0; i < this.boundMacaroons.length; i++) {
-      var boundMacaroon = this.boundMacaroons[i];
+    for (let i = 0; i < this.boundMacaroons.length; i++) {
+      const boundMacaroon = this.boundMacaroons[i];
       if (identifier === boundMacaroon.identifier) {
         return boundMacaroon;
       }
     }
-    return null;
+
+    return undefined;
   }
 
   private verifiesGeneral(caveat:string):boolean {
-    var found:boolean = false;
+    let found:boolean = false;
     for (var i = 0; i < this.generalCaveatVerifiers.length; i++) {
-      var verifier:GeneralCaveatVerifier = this.generalCaveatVerifiers[i];
+     const verifier:IGeneralCaveatVerifier = this.generalCaveatVerifiers[i];
       found = found || verifier(caveat);
     }
-    return found;
-  }
 
-  private static containsElement(elements:string[], anElement:string):boolean {
-    if (elements != null) {
-      for (var i = 0; i < elements.length; i++) {
-        var element = elements[i];
-        if (element === anElement) return true;
-      }
-    }
-    return false;
+    return found;
   }
 
 }
 
 
 class VerificationResult {
-  csig:Buffer = null;
-  fail:boolean = false;
-  failMessage:string = null;
+  public csig:Buffer = null;
+  public fail:boolean = false;
+  public failMessage:string = null;
 
-  constructor(csig:Buffer);
-  constructor(failMessage:string);
-  constructor(arg:any) {
+  constructor(arg:Buffer|string) {
     if (typeof arg === 'string') {
       this.failMessage = arg;
       this.fail = true;
@@ -261,7 +252,7 @@ class VerificationResult {
 }
 
 
-interface GeneralCaveatVerifier {
+interface IGeneralCaveatVerifier {
   /**
    * @param caveat caveat
    * @return True, if this caveat is satisfies the applications requirements. False otherwise.
