@@ -16,17 +16,18 @@
 
 /// <reference path="../../typings/tsd.d.ts" />
 
-export = BufferTools;
-class BufferTools {
 
-  public static equals(a:Buffer, b:Buffer):boolean {
-    if (!Buffer.isBuffer(a)) return undefined;
-    if (!Buffer.isBuffer(b)) return undefined;
-    if (typeof a['equals'] === 'function') return a['equals'](b);
-    if (a.length !== b.length) return false;
-    for (var i = 0; i < a.length; i++) {
-      if (a[i] !== b[i]) return false;
+export default class BufferTools {
+
+ public static equals(a:Buffer, b:Buffer):boolean {
+    if (!Buffer.isBuffer(a)) { return undefined; }
+    if (!Buffer.isBuffer(b)) { return undefined; }
+    if (typeof a['equals'] === 'function') { return a['equals'](b); }
+    if (a.length !== b.length) { return false; }
+    for (let i:number = 0; i < a.length; ) {
+      if (a[i] !== b[i]) { return false; }
     }
+
     return true;
   }
 
