@@ -16,30 +16,31 @@
 
 /// <reference path="../../typings/tsd.d.ts" />
 
-import Macaroon = require('./Macaroon');
-import CaveatPacket = require('./CaveatPacket');
-import CaveatPacketType = require('./CaveatPacketType');
-import MacaroonsConstants = require('./MacaroonsConstants');
-import Base64Tools = require('./Base64Tools');
 
-export = MacaroonsSerializer;
-class MacaroonsSerializer {
+import Base64Tools from './Base64Tools';
+import CaveatPacket from './CaveatPacket';
+import {CaveatPacketType} from './CaveatPacketType';
+import Macaroon from './Macaroon';
+import MacaroonsConstants from './MacaroonsConstants';
 
-  "use strict";
 
-  private static HEX = [0x30, 0x31, 0x32, 0x33, 0x34, 0x35, 0x36, 0x37, 0x38, 0x39, 0x61, 0x62, 0x63, 0x64, 0x65, 0x66];
+
+export default class MacaroonsSerializer {
+
+  private static HEX:number[] = [0x30, 0x31, 0x32, 0x33, 0x34, 0x35, 0x36, 0x37, 0x38, 0x39, 0x61, 0x62, 0x63, 0x64, 0x65, 0x66];
 
   public static serialize(macaroon:Macaroon):string {
-    var packets:Array<Buffer> = [];
+    let packets:Buffer[] = [];
     packets.push(
         MacaroonsSerializer.serialize_packet(CaveatPacketType.location, macaroon.location),
         MacaroonsSerializer.serialize_packet(CaveatPacketType.identifier, macaroon.identifier)
     );
-    packets = packets.concat(macaroon.caveatPackets.map(function (caveatPacket) {
+    packets = packets.concat(macaroon.caveatPackets.map((caveatPacket:CaveatPacket) => {
       return MacaroonsSerializer.serialize_packet_buf(caveatPacket.type, caveatPacket.rawValue)
     }));
     packets.push(MacaroonsSerializer.serialize_packet_buf(CaveatPacketType.signature, macaroon.signatureBuffer));
-    var base64 = MacaroonsSerializer.flattenByteArray(packets).toString("base64");
+    const base64 = MacaroonsSerializer.flattenByteArray(packets).toString("base64");
+    
     return Base64Tools.encodeBase64UrlSafe(base64);
   }
 
@@ -48,13 +49,13 @@ class MacaroonsSerializer {
   }
 
   private static serialize_packet_buf(type:CaveatPacketType, data:Buffer):Buffer {
-    var typname = CaveatPacketType[type];
-    var packet_len = MacaroonsConstants.PACKET_PREFIX_LENGTH + typname.length + MacaroonsConstants.KEY_VALUE_SEPARATOR_LEN + data.length + MacaroonsConstants.LINE_SEPARATOR_LEN;
-    var packet = new Buffer(packet_len);
+    const typname = CaveatPacketType[type];
+    const packetLength = MacaroonsConstants.PACKET_PREFIX_LENGTH + typname.length + MacaroonsConstants.KEY_VALUE_SEPARATOR_LEN + data.length + MacaroonsConstants.LINE_SEPARATOR_LEN;
+    const packet = new Buffer(packetLength);
     packet.fill(0);
-    var offset = 0;
+    let offset = 0;
 
-    MacaroonsSerializer.packet_header(packet_len).copy(packet, 0, 0);
+    MacaroonsSerializer.packet_header(packetLength).copy(packet, 0, 0);
     offset += MacaroonsConstants.PACKET_PREFIX_LENGTH;
 
     new Buffer(typname, 'ascii').copy(packet, offset, 0);
@@ -67,31 +68,38 @@ class MacaroonsSerializer {
     offset += data.length;
 
     packet[offset] = MacaroonsConstants.LINE_SEPARATOR;
+    
     return packet;
   }
 
   private static packet_header(size:number):Buffer {
     // assert.ok(size < 65536, "size < 65536");
-    var size = (size & 0xffff);
-    var packet = new Buffer(MacaroonsConstants.PACKET_PREFIX_LENGTH);
+    size = (size & 0xffff);
+    const packet = new Buffer(MacaroonsConstants.PACKET_PREFIX_LENGTH);
+    // tslint:disable-next-line:no-bitwise
     packet[0] = MacaroonsSerializer.HEX[(size >> 12) & 15];
+    // tslint:disable-next-line:no-bitwise
     packet[1] = MacaroonsSerializer.HEX[(size >> 8) & 15];
+    // tslint:disable-next-line:no-bitwise
     packet[2] = MacaroonsSerializer.HEX[(size >> 4) & 15];
+    // tslint:disable-next-line:no-bitwise
     packet[3] = MacaroonsSerializer.HEX[(size) & 15];
+    
     return packet;
   }
 
   private static flattenByteArray(bufs:Buffer[]):Buffer {
-    var size = 0;
-    for (var i = 0; i < bufs.length; i++) {
+    let size = 0;
+    for (let i = 0; i < bufs.length; i++) {
       size += bufs[i].length;
     }
-    var result = new Buffer(size);
+    let result = new Buffer(size);
     size = 0;
-    for (i = 0; i < bufs.length; i++) {
+    for (let i = 0; i < bufs.length; i++) {
       bufs[i].copy(result, size, 0);
       size += bufs[i].length;
     }
+    
     return result
   }
 
