@@ -14,8 +14,6 @@
  * limitations under the License.
  */
 
-export = TimestampCaveatVerifier;
-
 /**
  * <p>
  * A verifier that is able to verify timestamps against current time.
@@ -57,10 +55,10 @@ export = TimestampCaveatVerifier;
  * }</pre>
  */
 
-var CAVEAT_PREFIX = /time < .*/;
-var CAVEAT_PREFIX_LEN = "time < ".length;
+const CAVEAT_PREFIX = /time < .*/;
+const CAVEAT_PREFIX_LEN = "time < ".length;
 
-function TimestampCaveatVerifier(caveat:string):boolean {
+export default function TimestampCaveatVerifier(caveat:string):boolean {
   if (CAVEAT_PREFIX.test(caveat)) {
     var parsedTimestamp = new Date(caveat.substr(CAVEAT_PREFIX_LEN).trim());
     var time = parsedTimestamp.getTime();
