@@ -1,3 +1,3 @@
-/// <reference path="expect.js/expect.js.d.ts" />
+/// <reference path="chai/chai.d.ts" />
 /// <reference path="mocha/mocha.d.ts" />
 /// <reference path="node/node.d.ts" />
